Refactor data fetching
Remove bit of the duplication.
