+++ src/pages/EditExceptionOpeningHoursPage.tsx
@@ -1,12 +1,10 @@
-import React, { useEffect, useState } from 'react';
+import React from 'react';
 import api from '../common/utils/api/api';
-import {
-  ApiDatePeriod,
-  UiDatePeriodConfig,
-  Resource,
-} from '../common/lib/types';
-import { getDatePeriodFormConfig } from '../services/datePeriodFormConfig';
+import { ApiDatePeriod } from '../common/lib/types';
 import ExceptionOpeningHoursForm from '../components/exception-opening-hours-form/ExceptionOpeningHoursForm';
+import useResource from '../services/useResource';
+import useDatePeriodConfig from '../services/useDatePeriodConfig';
+import useDatePeriod from '../services/useDatePeriod';
 
 const EditExceptionOpeningHoursPage = ({
   resourceId,
@@ -15,41 +13,14 @@ const EditExceptionOpeningHoursPage = ({
   resourceId: string;
   datePeriodId: string;
 }): JSX.Element => {
-  const id = parseInt(datePeriodId, 10);
-  const [resource, setResource] = useState<Resource>();
-  const [datePeriodConfig, setDatePeriodConfig] = useState<
-    UiDatePeriodConfig
-  >();
-  const [datePeriod, setDatePeriod] = useState<ApiDatePeriod>();
+  const resource = useResource(resourceId);
+  const datePeriodConfig = useDatePeriodConfig();
+  const datePeriod = useDatePeriod(datePeriodId);
 
   const submitFn = (updatedDatePeriod: ApiDatePeriod): Promise<ApiDatePeriod> =>
     api.patchDatePeriod(updatedDatePeriod);
 
-  useEffect((): void => {
-    const fetchData = async (): Promise<void> => {
-      try {
-        const [
-          apiResource,
-          apiDatePeriod,
-          uiDatePeriodOptions,
-        ] = await Promise.all([
-          api.getResource(resourceId),
-          api.getDatePeriod(id),
-          getDatePeriodFormConfig(),
-        ]);
-        setResource(apiResource);
-        setDatePeriod(apiDatePeriod);
-        setDatePeriodConfig(uiDatePeriodOptions);
-      } catch (e) {
-        // eslint-disable-next-line no-console
-        console.error('Edit date-period - data initialization error:', e);
-      }
-    };
-
-    fetchData();
-  }, [id, resourceId]);
-
-  if (!resource || !datePeriodConfig) {
+  if (!resource || !datePeriodConfig || !datePeriod) {
     return <h1>Ladataan...</h1>;
   }
